feat: refactor this a bit
https://www.loom.com/share/7dd1b23e4b414ae69020ccdadec5744e?sid=fd756eab-f184-4345-8592-e35ef43d5b3b
