+++ apps/course-builder-web/src/lib/resources.ts
@@ -9,8 +9,5 @@ export async function getResource<T = any>(slugOrId: string): Promise<T | null> 
           summary,
           body,
           "slug": slug.current,
-          srt,
-          wordLevelSrt,
-          muxPlaybackId,
   }`)
 }
